development/requirements.txt
requests==2.25.1
Flask==2.0.1
PyJWT==2.3.0
